Move LocalPlatform into its own package.
For better public API.
