+++ core/src/main/scala/com/criteo/cuttle/local/package.scala
@@ -1,0 +1,10 @@
+package com.criteo.cuttle
+
+package object local {
+
+  implicit class InlineCommands(val sc: StringContext) extends AnyVal {
+    def sh(args: Any*) =
+      LocalPlatform.fork(sc.parts.zipAll(args, "", "").map { case (a, b) => a + b }.mkString)
+  }
+
+}
