+++ core/src/main/scala/com/criteo/cuttle/package.scala
@@ -11,9 +11,4 @@ package object cuttle {
 
   type SideEffect[S <: Scheduling] = (Execution[S]) => Future[Unit]
 
-  implicit class InlineCommands(val sc: StringContext) extends AnyVal {
-    def sh(args: Any*) =
-      LocalPlatform.fork(sc.parts.zipAll(args, "", "").map { case (a, b) => a + b }.mkString)
-  }
-
 }
